Added MS to HDFITS conversion
ms2hdf.py will take a MS and convert it into a HDFITS file. At the moment,
there is no attribute that shows this came from a MS, and not a FITS file.
The main issue that may arise is that conversion from MS -> HDF -> FITS will
be problematic, as keywords will be longer than 8 characters.