Documentation update -> pip installation
As of later version of python, six is not always installed by default. For this reason, I have added it to the setup file. 
It is now also possible to install the package directly without cloning the repository using pip. I have modified the documentation to notify this possibility, since it was already stated that this package might become available on pip at some point.

Thank you for the great software!